requirements/development.txt
asgiref==3.7.2
autopep8==2.0.4
Django==4.2.6
django-autoslug==1.9.9
django-dotenv==1.4.2
django-filter==23.3
django-silk==5.0.4
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gprof2dot==2022.7.29
Pillow==10.0.1
psycopg2-binary==2.9.9
pycodestyle==2.11.0
PyJWT==2.8.0
python-magic==0.4.27
pytz==2023.3.post1
sqlparse==0.4.4
Faker==19.9.0